Minimum required refactor
